requirements.txt
# ═══════════════════════════════════════════════════════════════════════════
# requirements.txt
# Multi-Criteria Recommender System (MCRS) — Full dependency list
#
# Install all:
#     pip install -r requirements.txt
#
# Python version: 3.9+
# ═══════════════════════════════════════════════════════════════════════════

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.135.3
uvicorn==0.44.0
starlette==1.0.0
python-multipart==0.0.26

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.13.0

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==5.0.0
cryptography==46.0.6

# ── Database ──────────────────────────────────────────────────────────────────
pymysql==1.1.2

# ── Scientific Computing ──────────────────────────────────────────────────────
numpy==2.4.3
pandas==3.0.1
scipy==1.17.1
scikit-learn==1.8.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.1
anyio==4.13.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.2.2

# ── JWT Dependencies ──────────────────────────────────────────────────────────
ecdsa==0.19.2
pyasn1==0.6.3
rsa==4.9.1
six==1.17.0